Move getMetadata to common.js
